pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-PharmaDDIEnv"
version = "0.1.0"
description = "PharmaDDIEnv environment for OpenEnv"
requires-python = ">=3.11"

dependencies = [
    "openenv-core[core]==0.2.3",
    "fastapi>=0.115.2",
    "uvicorn>=0.35.0",
    "pydantic>=2.7.0",
    "openai==2.30.0",
    "requests==2.32.3",
    "python-multipart==0.0.20",
    "python-dotenv==1.2.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m PharmaDDIEnv.server.app
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
package-dir = { "" = "." }

################################################################
